chore: remove unrelated cli/docs/test changes from refactor commit
